Move related state and functions down into SearchResultsPage
